Cody: add access token callback pages for CLI, Eclipse and VS (#63849)
Fixes CODY-2661

Previously, only the JetBrains and Neovim Cody plugins had dedicated
pages to authorize access token redirects to a localhost port. Other
clients like the Cody CLI and Eclipse plugin reused the JetBrains page,
which technically works but is problematic because the UI says the user
is authorizing "JetBrains" when in reality they're authorizing other
clients. This PR fixes the problem by adding dedicated pages for the
Cody CLI, Eclipse plugin and Visual Studio extension.

<!-- PR description tips:
https://www.notion.so/sourcegraph/Write-a-good-pull-request-description-610a7fd3e613496eb76f450db5a49b6e
-->

## Test plan

Didn't test this locally because it takes me a long time to setup Docker
access. The diff is small and should be low-risk to merge. I manually
searched for 'JETBRAINS' in this directory and confirmed there are no
remaining special cases for JetBrains.

<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->