Cody: add support for neovim, cody-cli, eclipse, and visualstudio clients
Fixes CODY-2884

Previously, the server could respond with a 406 error when encountering
clients with the following names: cody-cli, eclipse, and visualstudio.
This error seems to only happen for enterprise clients.  This PR fixes
the problem by adding special cases for these clients like we already
have for jetbrains and cody web.