fix(appliance): do not delete unowned resources (#63826)
The appliance checks owner references, and only deletes (or updates)
resources whose owner references matches the configmap being reconciled.

The current user interface to external databases, is for the admin to
create secrets with a well-known name out of band (e.g. "pgsql-auth")
and then disable the relevant backing services (e.g. pgsql). This commit
fixes a bug in which the appliance would have deleted such secrets,
leaving the admin unable to use external databases.

Discovered while investigating

https://linear.app/sourcegraph/issue/REL-14/ensure-pods-roll-when-referenced-secret-changes,
but does not close it.