feat(svelte): Improve small screeen behavior
This is an initial attempty to improve the web app for small screens.
This commit makes the following changes for small screens:

- Search home page:
  - No more search input overflow
  - History button is rendered on same line as other action buttons
    (saves some vertical space)
- Search results page:
  - Search input is rendered in the page, not in the header (more space)
  - Filters sidebar is hidden by default and can be shown via a
    `Filters` button.
  - The filters sidebar opens fullscreen and has a close button
  - The progrss button is smaller due to showing less information
- Repository pages:
  - File sidebar is hidden by default. It can be shown via a new button
    that is visible in the file headers
  - The file sidebar opens fullscreen and has a close button
  - NOTE: Selecting a file currently does not close the file sidebar,
    navigation happens in the background
  - Cody sidebar opens fullscreen and has the same close button
- General:
  - Fuzzy finder opens fullscreen and has a larger close button
  - Tabs don't show keyboard shortcuts

I tried to stick to CSS as much as possible but for some things to work
I had to change component structures or rendered elements conditionally.
Specifically when a component was already using `$isViewportMobile` I
usually just rendered elements conditionally.

I extended the `Panel` component to have a special 'mobile' mode, since
I realized I was doing similar changes to the filters, file tree and
cody sidebar.
The cody sidebar is a bit of a special case though because it's not
even rendered by default. So there are some additional steps required to
sync the open state.