feat(svelte): Add compare page (#63850)
This commit adds the Svelte version of the compare page. Now that we
have a revision picker this was easy to implement.

A couple of notes:
- The URL parameter is a rest parameter ([...spec]) because revisions
can contain slashes, and we don't encode those.
- I changed the API of the revision selector so that it works on pages
that don't have `resolvedRevision` available (also we never needed the
`repo` property from that object), and to provide custom select
handlers.
- Moved the revision to lib since it's used on various different pages.
- Unlike the React version this version
  - Provides forward/backward buttons to browse the commit list
  - Uses infinity scrolling to load the next diffs
- Doesn't use a different style for the commit list. I experimented with
adding a "compact" version of the commit but it didn't feel quite right.
I'm sure we'll eventually redesign this.
- The `filePath` parameter is supported (for backwards compatibility)
but we don't use it anywhere atm.

(note that the avatar size has changed since the video was recorded;
it's larger now)


## Test plan

Manual testing.