refactor(codeintel): Extracts a MappedIndex abstraction over uploads (#63781)
Precursor for
https://linear.app/sourcegraph/issue/GRAPH-735/test-syntactic-usages

This PR introduces the `MappedIndex` abstraction, which wraps up an
upload with a target commit. Its APIs then take care of mapping upload
relative paths and repo relative paths, and ranges across commits.

My main motivation for making this change is that I can now test this
logic in isolation (which this PR does), and also have an interface that
is easy to fake and use to test syntactic usages.

## Test plan

Added unit tests for the `MappedIndex` component, manual testing of the
GraphQL APIs show no changes in the syntactic usages output between this
PR and master.