feat(appliance): roll frontend deployment when pgsql secret changes
Ensure configHash changes by adding connection secret to the ifChanged
object, which causes createOrUpdateObject() in kubernetes.go to update
the object. Set the "checksum/auth" annotation on the pod template
in frontend's deployment to a hash of this connection object, to cause a
rolling deployment.

This annotation exists in helm, but this mechanism is slightly more
general as it always hashes secret values, whereas helm will sometimes
be solely hashing the name of a secret whose values may have changed.