show "Copied" indicator when copying an access token (#63936)
For some reason, the indication that the text was copied was removed for
secrets. The indication does not show the secret value itself, so I
don't see why that change was made. It is helpful to the user to give
them feedback that the secret was in fact copied.


## Test plan

Click the copy button after creating an access token