show org name not displayName in most places (#63907)
Previously, organizations would be displayed using their `displayName`,
which was inconsistent and non-canonical. The `displayName` is also
longer, so it's not suited to display in select menus and other similar
places.

For example, when viewing the user menu or creating a batch change or
saved search, the dropdown will now show `sourcegraph` instead of
`Sourcegraph` (where the latter is the display name).


![image](https://github.com/user-attachments/assets/79bce594-49c9-4603-9757-2e6785b26be9)


## Test plan

Go to create a batch change or saved search and confirm that org names
not displayNames are shown.

## Changelog

- The user menu and other filter menus now show the names of
organizations, not their "display names", to avoid ambiguity. For
example, these menus will now show `abc-corp` not `ABC Corp` (if the
latter was the `abc-corp`'s display name).