feat(appliance): roll frontend deployment when pgsql secret changes (#63845)
This approach is a bit of a departure from how our helm chart works.
That hashes hashes values blocks to determine these annotations. The
issue is, that these values blocks often contain references to secrets,
and therefore don't change when the _content_ of the secret changes.

This PR takes a more general approach of looking for that secret, which
must always eventually exist - whether admin-provisioned or provisioned
by the pgsql service definition - and hashes that.