Fixes re: Cody Pro x Sourcegraph-supplied models (#63892)
This PR fixes several issues related to CodyPro/Sourcegraph.com relying
on Sourcegraph-supplied models.

When Sourcegraph.com is _exclusively_ using the modelconfig systems we
can delete a lot of this code, and all of the hard-coded lists of LLM
models. However, in order to support a smoother transition, we need to
support Sourcegraph.com _not_ using Sourcegraph-supplied models, as well
as updating the site configuration so that it does.

This PR makes the following fixes:

- For forwards compatibility, Sourcegraph.com will now support model
references using the "mref format" (a::b::c). Whereas previously it
_only_ would support the legacy format (a/b). This will allow us to
update Cody clients to use mrefs natively.

- Add virtualized models to the hard-coded list of Cody Pro chat models.
If Sourcegraph.com is using Sourcegraph-supplied LLM models, then it
will use ModelID "claude-3-sonnet" to refer to the LLM model with
ModelName "claude-3-sonnet-20240229". We now accept the virtualized
model name inside of `func isAllowedCodyProChatModel(...)`, as well as
"devirtualize" the model name inside of `func
resolveRequestedModel(..)`.

- Use the Cody Pro user's access token for calls to Cody Gateway. This
is a serious bug that is live today. (⚠️😬) For Sourcegraph.com, when we
call Cody Gateway we do _not_ want to authenticate the request using the
Sourcegraph license key. And instead use the end user's credentials.
However, inside of the Cody Gateway completions client, we were not
actually using that access token.

See:
https://github.com/sourcegraph/sourcegraph/blob/dbf420bb8c359c71c07934b983116bdbb8f41b06/internal/completions/types/types.go#L61-L75

> I'll try to add a unit test to ensure this behavior before submitting,
but wanted to get the PR sent out for review ASAP.

## Test plan

Added, updated unit tests. Tested manually.

## Changelog

NA