chore(appliance): make a function less misleading (#63994)
isNamespaced was implemented the opposite way to its name, returning
true when a resource was _not_ namespaced. We could have renamed it to
isNotNamespaced, but then most callsites would look like a double
negative. So this commit inverts the implementation to match the name,
and also its callers.

<!-- PR description tips:
https://www.notion.so/sourcegraph/Write-a-good-pull-request-description-610a7fd3e613496eb76f450db5a49b6e
-->

Forked from https://github.com/sourcegraph/sourcegraph/pull/63972 to
clean up the branch

## Test plan

<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->
CI

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

Co-authored-by: Craig Furman <craig.furman@sourcegraph.com>